wip test refactor
